FIXED some build errors
