+++ prisma/schema.prisma
@@ -91,7 +91,6 @@ model NavigationBarLink {
   id    String @id @default(auto()) @map("_id") @db.ObjectId
   title String
   url   String
-  )
 }
 
 // Address is an embedded document
